move UiPrinter to ui.js
