+++ packages/mockotlpserver/lib/printers.js
@@ -7,9 +7,6 @@
  *      printer.subscribe();
  */
 
-const fs = require('fs');
-const path = require('path');
-
 const {
     diagchSub,
     CH_OTLP_V1_LOGS,
@@ -81,66 +78,7 @@ class InspectPrinter extends Printer {
     }
 }
 
-/**
- * @typedef {Object} SpanTree
- * @property {import('./types').Span} span
- * @property {import('./types').Span[]} children
- */
-
-/**
- * @typedef {Object} TraceTree
- * @property {string} id
- * @property {SpanTree[]} children
- */
-
-class UiPrinter extends Printer {
-    constructor(log) {
-        super(log);
-        this._dbDir = path.resolve(__dirname, '../db');
-    }
-
-    /**
-     * Prints into files the spns belonging to a trace
-     * @param {import('./types').ExportTraceServiceRequest} traceReq
-     */
-    printTrace(traceReq) {
-        /** @type {Map<string, import('./types').Span[]>} */
-        const tracesMap = new Map();
-
-        // Group all spans by trace
-        traceReq.resourceSpans.forEach((resSpan) => {
-            resSpan.scopeSpans.forEach((scopeSpan) => {
-                scopeSpan.spans.forEach((span) => {
-                    const traceId = span.traceId.toString('hex');
-                    let traceSpans = tracesMap.get(traceId);
-
-                    if (!traceSpans) {
-                        traceSpans = [];
-                        tracesMap.set(traceId, traceSpans);
-                    }
-                    traceSpans.push(span);
-                });
-            });
-        });
-
-        // Write into a file
-        // TODO: manage lifetime of old trace ndjson files.
-        for (const [traceId, traceSpans] of tracesMap.entries()) {
-            const filePath = path.join(this._dbDir, `trace-${traceId}.ndjson`);
-            const stream = fs.createWriteStream(filePath, {
-                flags: 'a',
-                encoding: 'utf-8',
-            });
-
-            for (const span of traceSpans) {
-                stream.write(JSON.stringify(span) + '\n');
-            }
-            stream.close();
-        }
-    }
-}
-
 module.exports = {
+    Printer,
     InspectPrinter,
-    UiPrinter,
 };
